+++ src/App.css
@@ -20,3 +20,9 @@
   color: #61dafb;
 }
 
+.page{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
