+++ src/components/Certificate.css
@@ -1,0 +1,114 @@
+.certificate-container{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-bottom: 5%;
+    width: 80%;
+}
+
+.certificate-title{
+    margin: 3%;
+}
+
+
+.certificate-box{
+    position: relative;
+    width: 100%;
+    height: 300px;
+    margin: 4%;
+}
+
+.certificate{
+    display: flex;
+    flex-direction: column;
+    gap:10px;
+    justify-content: center;
+    align-items: center;
+    position: absolute;
+    width: 300px;
+    height: auto;
+    padding: 1%;
+    background-color: #090116;
+    top: 50%;
+    left:50%;
+    border: 2px solid rgb(49, 11, 119);
+    border-radius: 20px;
+    transition: transform 0.5s ease-in-out;
+}
+
+.certificate.one{
+    height: 100%;
+    transform: translate(-50%, -50%) rotate(10deg);  
+}
+.certificate img{
+    height: auto;
+    width: 300px;
+}
+.certificate.two{
+    height: 100%;
+    transform: translate(-55%, -45%) rotate(5deg);  
+}
+
+.certificate.three{
+    height: 100%;
+    transform: translate(-60%, -40%) rotate(0deg);
+}
+
+.certificate-box:hover .one{
+    transform: translate(60%, -50%) rotate(0deg);
+}
+
+.certificate-box:hover .two{
+    transform: translate(-50%, -50%) rotate(0deg);
+}
+
+.certificate-box:hover .three{
+    transform: translate(-160%, -50%) rotate(0deg);
+}
+
+.image-overlay{
+    position: fixed;
+    top:0;
+    right:0;
+    left:0;
+    bottom: 0;
+    background: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 2;
+}
+
+.image-content{
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    overflow: hidden;
+    background: rgba(151, 151, 151, 0.5);
+    padding: 20px;
+    border-radius: 15px;
+    max-width: 80%;
+    max-height: 80%;
+}
+
+.image-content img{
+    width: 50%;
+    height: 50%;
+}
+
+.image-content p{
+    text-align: center;
+    font-size: 2em;
+    font-weight: bolder;
+}
+
+.close-image{
+    position: absolute;
+    top:10px;
+    right:10px;
+    cursor: pointer;
+    font-size: 24px;
+    font-weight: bold;
+}
