+++ src/components/Experience.css
@@ -1,0 +1,94 @@
+.experience-container{
+    display: flex;
+    flex-direction: column;
+    border:solid 1px white;
+    width: 60%;
+}
+
+.experience{
+    align-items: center;
+    display: flex;
+    justify-content:flex-start;
+    font-size: 1.5em;
+    margin: 1%;
+}
+
+
+.jobs-container{
+    display: flex;
+    justify-content: center;
+    gap:2%;
+    margin:2%;
+    border:1px solid white
+}
+
+.button-tile{
+    display: flex;
+    flex-direction: column;
+    width: 18%;
+}
+
+.job{
+    color:rgb(202, 171, 255);
+    background-color: transparent;
+    padding: 10%;
+    font-size: 1rem;
+    text-align: start;
+    border:none;
+    border-left:3px solid rgb(33, 0, 89);
+    cursor: pointer;
+    
+}
+
+.job:hover{
+    background-color: rgb(30, 2, 62);
+    color:rgb(201, 144, 255);
+}
+
+.job[aria-selected = "true"]{
+    color:rgb(201, 144, 255);
+    transition: border-left 0.3s ease-in-out;
+    border-left:3px solid rgb(183, 0, 255); 
+    background-color: rgb(30, 2, 62);
+    animation: borderMoveUpDown 0.3s ease-in-out forwards;
+}
+@keyframes borderMoveUpDown {
+    0% {
+        transform: translateY(0);
+    }
+    50% {
+        transform: translateY(-10px);
+    }
+    100% {
+        transform: translateY(0);
+    }
+}
+
+.job-card{
+    width: 60%;
+}
+
+.job-title{
+    font-size: 23px
+}
+
+.job-time{
+    font-size: 1em;
+    margin:2%;
+}
+
+.arrow-text-container{
+    position: relative;
+    display: flex;
+    gap:1%;
+    margin: 4%;
+}
+.arrow {
+    position: relative;
+    width: 0;
+    height: 0;
+    top:7px;
+    border-top: 4px solid transparent;
+    border-bottom: 4px solid transparent;
+    border-left: 6px solid rgb(93, 0, 255);
+}
