+++ src/components/Project.css
@@ -1,0 +1,159 @@
+.project{
+    align-items: center;
+    display: flex;
+    justify-content:flex-start;
+    font-size: 1.5em;
+    margin: 2%;
+}
+
+.project-container{
+    display: flex;
+    position: relative;
+    flex-direction: column;
+    margin: 40px;
+    width: 80%;
+}
+
+.project-card{
+    display: flex;
+    position: relative;
+    align-items: center;
+    justify-content: space-evenly;
+    margin: 4%;
+}
+
+.imgContainer{
+    position: relative;
+    display: flex;
+    width: 50%;
+    cursor: pointer;
+    align-items: center;
+    justify-content: center;
+}
+.project-card img {
+    width: 100%;
+    height: auto;
+    z-index: -1;
+}
+
+.video-overlay{
+    position: fixed;
+    top:0;
+    left:0;
+    right: 0;
+    bottom: 0;
+    background: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index:2;
+}
+
+.video-content{
+    background: #c5c5c538;
+    padding: 20px;
+    border-radius: 15px;
+    max-width: 80%;
+    max-height: 80%;
+    position: relative;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    overflow: hidden;
+}
+
+.video-content video{
+    width: 80%;
+    height: auto;
+}
+
+.close-button{
+    position: absolute;
+    top:10px;
+    right:10px;
+    cursor: pointer;
+    font-size: 24px;
+    font-weight: bold;
+} 
+
+.videoDemo{
+    position: absolute;
+    color:white;
+    background: rgba(0, 0, 0, 0.5);
+    bottom: 0;
+    width: 100%;
+    height: 20%;
+    font-size: 20px;
+    align-items: center;
+    justify-content: center;
+    display: flex;
+    opacity: 0;
+    transition: 0.3s;
+}
+
+.project-card:hover .videoDemo{
+    opacity: 1;
+}
+
+.project-info{
+    position: relative;
+    padding: 2%;
+    /* border: 1px solid white; */
+    display: flex;
+    justify-content: space-between;
+    flex-direction: column;
+    width: 50%;
+}
+
+.left-info{
+    align-items: start;
+}
+
+.right-info{
+    align-items: end;
+}
+.project-description{
+    position: relative;
+    padding: 2%;
+    display: flex;
+    margin: 1%;
+    padding: 3%;
+    align-items: center;
+    background-color: #352255;
+    width: 115%;
+    height: 100%;
+    border-radius:5px;
+}
+.softwares{
+    word-spacing: 2em;
+}
+.left-desc{
+    justify-content: left;
+}
+
+.right-desc{
+    justify-content: right;
+}
+
+.other-project-container{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-bottom: 10%;
+}
+
+.other-title{
+    margin: 3%;
+}
+
+.other-project-box{
+    display: grid;
+    justify-content: space-between;
+    grid-template-columns: 22vw 22vw 22vw;
+    height: 44vh;
+    column-gap: 2vw;
+}
+
+.box{
+    border:solid 1px white;
+}
